Warning removed.
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@687 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ libhxcfe/sources/loaders/cpcdsk_loader/cpcdsk_writer.c
@@ -189,7 +189,7 @@ int CPCDSK_libWrite_DiskFile(HXCFLOPPYEMULATOR* floppycontext,FLOPPY * floppy,ch
 						k=0;
 						do
 						{
-							if(sca[k]->sectorsize!=sectorsize)
+							if(sca[k]->sectorsize!=(unsigned int)sectorsize)
 							{
 								sectorsize=-1;
 							}
